requirements.txt
datasets==2.12.0
Pillow==9.5.0
Pillow==9.5.0
streamlit==1.23.1
torch==2.0.1+cu117
torchmetrics==0.11.4
torchvision==0.15.2+cu117
tqdm==4.65.0
transformers==4.29.2
